Moved reduction kernels into a header file
